+++ src/classifiers.js
@@ -4,7 +4,7 @@
  * @module classifiers
  */
 
-const {hasComponentTag, hasHOCTag, hasStaticTag} = require('./utils');
+const {hasComponentTag, hasHOCTag} = require('./utils');
 
 function defaultTypeClassifier ({section}) {
 	const kind = section.kind;
@@ -12,11 +12,6 @@ function defaultTypeClassifier ({section}) {
 	// Check for HOC
 	if (hasHOCTag(section)) {
 		return 'hoc';
-	}
-
-	// Check for @static
-	if (hasStaticTag(section)) {
-		return 'static';
 	}
 
 	// Check for @ui